Extract input stuff
